+++ examples/Unity/Expedition/ServerSolution/Expedition.Server/Program.cs
@@ -114,8 +114,9 @@ builder.Services.AddSignalR(hubOptions =>
 // Authentication
 builder.Services.AddMetaAuth(options =>
 {
-    options.SecretKey = "expedition-secret-key-at-least-32-characters!!";
+    options.SecretKey = "expedition-secret-key-at-least-32-characters!!_new3";
     options.Issuer = "expedition-server";
+    options.AccessTokenLifetime = TimeSpan.FromMinutes(20);
 });
 builder.Services.AddSingleton(new MetaTransportOptions
 {